update from ft_ls

+++ src/lib/ft_atoi.c
@@ -6,13 +6,13 @@
 /*   By: alucas- <alucas-@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/07 09:52:30 by alucas-           #+#    #+#             */
-/*   Updated: 2017/11/15 18:30:38 by null             ###   ########.fr       */
+/*   Updated: 2017/11/17 09:57:36 by null             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "libft/lib.h"
 
-inline int	ft_atoi(char const *str)
+inline int64_t	ft_atoi(char const *str)
 {
 	int64_t	result;
 	char	sign;
@@ -24,5 +24,5 @@ inline int	ft_atoi(char const *str)
 		++str;
 	while (ft_isdigit(*str))
 		result = result * 10 + *str++ - '0';
-	return ((int)(sign == '-' ? -result : result));
+	return ((int64_t)(sign == '-' ? -result : result));
 }
